--- src/main/java/org/prgrms/kdt/OrderRepository.java
@@ -1,5 +1,0 @@
-package org.prgrms.kdt;
-
-public interface OrderRepository {
-    void insert(Order order);
-}
