+++ src/main/java/org/prgrms/kdt/OrderTester.java
@@ -1,5 +1,7 @@
 package org.prgrms.kdt;
 
+import org.prgrms.kdt.entity.OrderItem;
+import org.prgrms.kdt.service.OrderService;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.util.Assert;
