+++ src/main/java/org/prgrms/kdt/entity/PercentDiscountVoucher.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdt;
+package org.prgrms.kdt.entity;
 
 import java.util.UUID;
