+++ src/main/java/org/prgrms/kdt/repository/OrderRepository.java
@@ -1,0 +1,7 @@
+package org.prgrms.kdt.repository;
+
+import org.prgrms.kdt.entity.Order;
+
+public interface OrderRepository {
+    void insert(Order order);
+}
