+++ src/main/java/org/prgrms/kdt/service/OrderService.java
@@ -1,4 +1,9 @@
-package org.prgrms.kdt;
+package org.prgrms.kdt.service;
+
+import org.prgrms.kdt.entity.Order;
+import org.prgrms.kdt.entity.OrderItem;
+import org.prgrms.kdt.repository.OrderRepository;
+import org.prgrms.kdt.entity.Voucher;
 
 import java.util.*;
